+++ DesignToolbox/DesignToolbox/Utils/Bundle+extension.swift
@@ -17,30 +17,33 @@ extension Bundle {
 
     // MARK: API
 
-    /// Marketing version in semantic versioning
+    /// Marketing version in semantic versioning, for example "0.15.0".
+    /// Mapped to field *CFBundleShortVersionString* in project settings.
     var marketingVersion: String {
         string(forInfoDictionaryKey: "CFBundleShortVersionString") ?? "0.0.0"
     }
 
-    /// Incremental build number
+    /// Incremental build number, for example timestamp or manualy defined value.
+    /// Mapped to value *kCFBundleVersionKey* in project settings.
     var buildNumber: String {
         string(forInfoDictionaryKey: kCFBundleVersionKey as String) ?? "-1"
     }
 
-    /// Type of build. Should be within:
-    /// - DEBUG for developers builds
-    /// - ALPHA for features tests CI/CD builds
-    /// - BETA for develop CI/CD builds
-    /// - PROD for main releases (CI/CD builds)
+    /// Type of build. Should be within (see Fastfile):
+    /// - "debug" for developers builds
+    /// - "alpha" for features tests (CI/CD builds)
+    /// - "beta" for develop branch (CI/CD builds nightly builds)
+    /// - "stable" for main releases (CI/CD builds)
     var buildType: String? {
         #if DEBUG
-        "DEBUG"
+        "debug"
         #else
         string(forInfoDictionaryKey: "OUDSBuildType")
         #endif
     }
 
-    /// A value filled outside the app (e.g. by the CI/CD chain) for debug purposes to help to match the current app codebase to some CI/CD build tag
+    /// A value filled outside the app (e.g. by the CI/CD chain) for debug purposes to help to match the current app codebase to some CI/CD build tag.
+    /// For example, tag suffix for CI and TestFlight Git tags with hash of commits.
     var buildTag: String? {
         #if DEBUG
         nil
@@ -49,12 +52,23 @@ extension Bundle {
         #endif
     }
 
-    /// A value filled outside the app (e.g. by the CI/CD chain) for debug purposes to help to match the current app codebase to some issues to test
+    /// A value filled outside the app (e.g. by the CI/CD chain) for debug purposes to help to match the current app codebase to some issues to test.
+    /// For example, contains the issues numbers.
     var buildDetails: String? {
         #if DEBUG
         nil
         #else
         string(forInfoDictionaryKey: "OUDSBuildDetails")
+        #endif
+    }
+
+    /// A value filled outside the app (e.g. by the CI/CD chain) to indicate the version of the OUDS iOS Swift Package in use
+    /// For example, contains a tag or a branch or a commit hash
+    var sdkVersion: String? {
+        #if DEBUG
+        nil
+        #else
+        string(forInfoDictionaryKey: "OUDSSDKVersion")
         #endif
     }
 
@@ -67,6 +81,7 @@ extension Bundle {
         return "\(type)\(tag)"
     }
 
+    // WARNING: Do not change this value below as automatically parsed by script
     var tokensLibraryVersion: String {
         "Tokens version: 0.11.0"
     }