+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- [DesignToolbox] Display of build details about OUDS iOS package version in about page (Orange-OpenSource/ouds-ios#679)
 - [DesignToolbox] Add title in Component and Token screens (Orange-OpenSource/ouds-ios#662)
 - [Tool] Update `SwiftFormat/CLI` pod from v0.55.5 to v0.56.1
 - [Tool] Update `fastlane` RubyGem from v2.227.1 to v2.227.2